requirements.txt
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
pandas==1.5.2
requests==2.28.1
openpyxl==3.0.10
sqlalchemy==1.4.44
databases==0.6.2
aiosqlite==0.17.0